+++ tests/util/mod.rs
@@ -16,7 +16,7 @@ lazy_static! {
     static ref LOCK: Arc<Mutex<()>> = Arc::new(Mutex::new(()));
 }
 
-pub fn create_backing_file(size: i64) -> TempPath {
+pub fn create_backing_file(size: i32) -> TempPath {
     let file = NamedTempFile::new().expect("should be able to create a temp file");
     if unsafe { fallocate(file.as_raw_fd(), 0, 0, size.try_into().unwrap()) } < 0 {
         panic!(
